Extract RichTextCell & use for Variance Explanation cells
